fix: move config to types
