+++ src/common/feedGenerator.ts
@@ -29,7 +29,7 @@ import graphorm from '../graphorm';
 import { mapArrayToOjbect } from './object';
 import { runInSpan } from '../telemetry';
 import { whereVordrFilter } from './vordr';
-import { baseFeedConfig } from '../integrations/feed/generators';
+import { baseFeedConfig } from '../integrations/feed';
 
 export const WATERCOOLER_ID = 'fd062672-63b7-4a10-87bd-96dcd10e9613';
